requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
boto3==1.29.3
pydantic==2.4.2
slowapi==0.1.8
python-multipart==0.0.6 